+++ src/schema/models/user.ts
@@ -1,6 +1,7 @@
 import { objectType, extendType, stringArg } from '@nexus/schema'
 import { GraphQlContext } from '../../../types'
 import { createPassword, verifyPassword } from '../../utils/password'
+import { sendEmailWithTemplate } from '../../../libs/Mailjet'
 import jwt from 'jsonwebtoken'
 
 export const User = objectType({
@@ -23,24 +24,33 @@ export const login = extendType({
         password: stringArg({ required: true }),
       },
       resolve: async (_: unknown, { email, password }, { prisma }: GraphQlContext) => {
+        let user
+
         try {
-          const user = await prisma.user.findOne({
+          user = await prisma.user.findOne({
             where: { email },
           })
-
-          if (!user) {
-            throw new Error('Invalid email or password')
-          }
-
-          const isMatchPassword = verifyPassword(password, user.password)
-          if (!isMatchPassword) {
-            throw new Error('Invalid email or password')
-          }
-
-          return jwt.sign(user, process.env.JWT_SECRET as string)
         } catch (error) {
           throw new Error(error.code)
         }
+
+        if (!user) {
+          throw new Error('Invalid email or password')
+        }
+
+        const isMatchPassword = await verifyPassword(password, user.password)
+        if (!isMatchPassword) {
+          throw new Error('Invalid email or password')
+        }
+
+        return jwt.sign(
+          {
+            id: user.id,
+            email: user.email,
+            name: user.name,
+          },
+          process.env.JWT_SECRET as string,
+        )
       },
     })
   },
@@ -58,13 +68,17 @@ export const register = extendType({
       },
       resolve: async (_: unknown, { email, password, name }, { prisma }: GraphQlContext) => {
         try {
-          return prisma.user.create({
+          const newUser = await prisma.user.create({
             data: {
               name,
               email,
               password: await createPassword(password),
             },
           })
+
+          await sendEmailWithTemplate([{ Email: email }], 1394707)
+
+          return newUser
         } catch (error) {
           throw new Error(error.code)
         }
@@ -78,13 +92,15 @@ export const deleteUser = extendType({
   definition(t): void {
     t.field('deleteUser', {
       type: 'User',
-      args: {
-        id: stringArg({ required: true }),
-      },
-      resolve: async (_: unknown, { id }, { prisma }: GraphQlContext) => {
+      args: {},
+      resolve: async (_: unknown, {}, { prisma, user }: GraphQlContext) => {
+        if (!user) {
+          throw new Error('Authentification required !')
+        }
+
         try {
           return prisma.user.delete({
-            where: { id },
+            where: { email: user.email },
           })
         } catch (error) {
           throw new Error(error.code)
@@ -100,18 +116,21 @@ export const updateUser = extendType({
     t.field('updateUser', {
       type: 'User',
       args: {
-        id: stringArg({ required: true }),
         name: stringArg({ required: false }),
         email: stringArg({ required: false }),
       },
-      resolve: async (_: unknown, { id, email, name }, { prisma }: GraphQlContext) => {
+      resolve: async (_: unknown, { email, name }, { prisma, user }: GraphQlContext) => {
+        if (!user) {
+          throw new Error('Authentification required !')
+        }
+
         try {
           return prisma.user.update({
             data: {
               email: email,
               name: name,
             },
-            where: { id },
+            where: { email: user.email },
           })
         } catch (error) {
           throw new Error(error.code)
